requirements.txt
google-genai==1.2.0
google-api-python-client==2.160.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
jsonschema==4.21.1
python-dotenv==1.0.1
rich==13.7.1
fastapi==0.111.0
uvicorn==0.30.0
httpx==0.27.0
pydantic>=2.0.0
pytest==8.0.0
